requirements.txt
asttokens==2.0.7
async-generator==1.10
attrs==19.3.0
Automat==0.8.0
backcall==0.2.0
beautifulsoup4==4.11.1
blinker==1.4
branca==0.5.0
bs4==0.0.1
certifi==2022.6.15
chardet==3.0.4
Click==7.0
click-plugins==1.1.1
cligj==0.7.2
cloud-init==20.1
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==2.8
cycler==0.11.0
dbus-python==1.2.16
debugpy==1.6.2
decorator==5.1.1
distro==1.4.0
distro-info==0.23ubuntu1
dython==0.7.2
entrypoints==0.3
et-xmlfile==1.1.0
executing==0.10.0
Fiona==1.8.21
folium==0.12.1.post1
fonttools==4.34.4
geographiclib==1.52
geopandas==0.11.1
geopy==2.2.0
git-filter-repo==2.34.0
glmnet==2.2.1
h11==0.13.0
httplib2==0.14.0
hyperlink==19.0.0
idna==2.8
imgkit==1.2.2
importlib-metadata==1.5.0
incremental==16.10.1
ipykernel==6.15.1
ipython==8.4.0
jedi==0.18.1
Jinja2==2.10.1
joblib==1.1.0
jsonmerge==1.8.0
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
jupyter-client==7.3.4
jupyter-core==4.11.1
keyring==18.0.1
kiwisolver==1.4.4
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
MarkupSafe==1.1.0
matplotlib==3.5.3
matplotlib-inline==0.1.3
more-itertools==4.2.0
munch==2.5.0
nest-asyncio==1.5.5
netifaces==0.10.4
numpy==1.23.2
oauthlib==3.1.0
openpyxl==3.0.10
openrouteservice==2.3.3
outcome==1.2.0
packaging==21.3
pandas==1.4.3
parso==0.8.3
patsy==0.5.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
pip==20.0.2
prompt-toolkit==3.0.30
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
py4j==0.10.9.5
pyarrow==9.0.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
Pygments==2.12.0
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==3.0.9
PyPDF2==2.10.9
pyproj==3.3.1
pyrsistent==0.15.5
pyserial==3.4
PySocks==1.7.1
pyspark==3.3.0
python-apt==2.0.0
python-dateutil==2.8.2
python-debian==0.1.36ubuntu1
python-dotenv==0.21.0
pytz==2022.2.1
PyYAML==5.3.1
pyzmq==23.2.1
requests==2.22.0
requests-unixsocket==0.2.0
scikit-learn==1.1.2
scikit-plot==0.3.7
scipy==1.9.0
seaborn==0.11.2
SecretStorage==2.3.1
selenium==4.4.3
service-identity==18.1.0
setuptools==45.2.0
Shapely==1.8.2
simplejson==3.16.0
six==1.14.0
sklearn==0.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
ssh-import-id==5.10
stack-data==0.4.0
statsmodels==0.13.2
systemd-python==234
threadpoolctl==3.1.0
tornado==6.2
tqdm==4.64.1
traitlets==5.3.0
trio==0.21.0
trio-websocket==0.9.2
Twisted==18.9.0
typing-extensions==4.3.0
ubuntu-advantage-tools==20.3
ufw==0.36
unattended-upgrades==0.1
uniblock==1.0.4
urllib3==1.26.11
wadllib==1.3.3
wcwidth==0.2.5
webdriver-manager==3.8.3
wheel==0.34.2
wsproto==1.1.0
xlrd==2.0.1
zipp==1.0.0
zope.interface==4.7.1
